avoid and ignore tput errors
tput tries to fetch TERM, which may not be set and causes it to fail.
using -Txterm sets the terminal to xterm, no matter what's currently used.
It will also cause tput to not query the screen size, which we don't need anyway.

Finally, we ignore any other errors possibly caused by tput with set +e